[REF] merge abstract fiscal documents and concrete fiscal documents
